Title: 합과 평균의 차, Time: 105ms, Memory: 24MB, Status: Passed - CodeTree

+++ 240116/합과 평균의 차/sub-of-average-and-sum.py
@@ -1,5 +1,5 @@
 a,b,c = map(int,input().split())
 t=a+b+c
 print(t)
-print(t/3)
-print(t-t/3)
+print(int(t/3))
+print(t-int(t/3))
